python/fastAPI/requirements.txt
fastapi==0.108.0
pydantic==2.5.2
pydantic-settings==2.2.1
uvicorn==0.25.0
aiofiles==23.1.0
pytest==7.1.3
requests==2.31.0

# UMA & OIDC
python-keycloak==3.7.0
Authlib==1.3.0
requests==2.31.0

# Debugger for VSC
debugpy==1.4.1
